2025.12.24
왕복, 편도 로직 구현
API호출시 DB에 데이터 INSERT-> INSERT한 데이터 조회

+++ BackEnd_TripType_Final_Project/src/main/java/com/kh/triptype/admin/pricing/model/vo/FlightSearchCacheVo.java
@@ -23,7 +23,7 @@ public class FlightSearchCacheVo {
     private int flightPriceHistoryNo;
 
     /** 항공편 제안 ID (외부 API 기준, 논리적 식별자) */
-    private int flightOfferId;
+    private long flightOfferId;
 
     /** 항공권 총 가격 */
     private BigDecimal flightOfferPriceTotal;
